tests/wpt/web-platform-tests/css/css-images/object-view-box-parsing.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS object-view-box: computed values</title>
<link rel="author" title="Khushal Sagar" href="mailto:khushalsagar@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-images-4/#the-object-view-box">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<div id=target></div>
<script>
test_valid_value('object-view-box', 'inset(10%)');
test_valid_value('object-view-box', 'inset(10px 20px 30px 40px)');
test_valid_value('object-view-box', 'rect(1px 10% 100px 50%)');
test_valid_value('object-view-box', 'rect(auto auto auto auto)');
test_valid_value('object-view-box', 'rect(auto 5px auto 10%)');
test_valid_value('object-view-box', 'xywh(1px 10% 100px 50%)');

test_invalid_value('object-view-box', 'circle(10px)');
test_invalid_value('object-view-box', 'ellipse(10px 20px)');
test_invalid_value('object-view-box', 'polygon(10px 20px 30px)');
test_invalid_value('object-view-box', 'path("M20,80 L50,20")');
test_invalid_value('object-view-box', 'blah');
test_invalid_value('object-view-box', 'xywh(1px 10%)');

</script>
</body>
</html>
